Progress fixing search results
